+++ src/ContaoManager/Plugin.php
@@ -47,10 +47,11 @@ class Plugin implements BundlePluginInterface, ExtensionPluginInterface, ConfigP
      */
     public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
     {
-        $loader->load('@HeimrichHannotContaoUtilsBundle/Resources/config/services.yml');
+        $loader->load('@HeimrichHannotContaoUtilsBundle/Resources/config/commands.yml');
         $loader->load('@HeimrichHannotContaoUtilsBundle/Resources/config/listener.yml');
         $loader->load('@HeimrichHannotContaoUtilsBundle/Resources/config/parameters.yml');
+        $loader->load('@HeimrichHannotContaoUtilsBundle/Resources/config/services.yml');
+        $loader->load('@HeimrichHannotContaoUtilsBundle/Resources/config/twig.yml');
         $loader->load('@HeimrichHannotContaoUtilsBundle/Resources/config/utils.yml');
-        $loader->load('@HeimrichHannotContaoUtilsBundle/Resources/config/commands.yml');
     }
 }