+++ src/Twig/StringExtension.php
@@ -8,14 +8,21 @@
 
 namespace HeimrichHannot\UtilsBundle\Twig;
 
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
-use Symfony\Component\DependencyInjection\ContainerAwareTrait;
+use HeimrichHannot\UtilsBundle\String\AnonymizerUtil;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFilter;
 
-class StringExtension extends AbstractExtension implements ContainerAwareInterface
+class StringExtension extends AbstractExtension
 {
-    use ContainerAwareTrait;
+    /**
+     * @var AnonymizerUtil
+     */
+    private $anonymizerUtil;
+
+    public function __construct(AnonymizerUtil $anonymizerUtil)
+    {
+        $this->anonymizerUtil = $anonymizerUtil;
+    }
 
     /**
      * Get list of twig filters.
@@ -26,6 +33,7 @@ class StringExtension extends AbstractExtension implements ContainerAwareInterfa
     {
         return [
             new TwigFilter('autolink', [$this, 'autolink']),
+            new TwigFilter('anonymize_email', [$this, 'anonymizeEmail']),
         ];
     }
 
@@ -38,5 +46,10 @@ class StringExtension extends AbstractExtension implements ContainerAwareInterfa
 
             return '<a'.(isset($options['blank']) && $options['blank'] ? ' target=" _blank"' : '').' href="'.$matches['url'].'">'.$matches['url'].'</a>';
         }, $text, $options['limit'] ?? -1);
+    }
+    
+    public function anonymizeEmail(string $text): string
+    {
+        return $this->anonymizerUtil->anonymizeEmail($text);
     }
 }